move middleware definition
