+++ openhands/server/shared.py
@@ -4,7 +4,7 @@ import socketio
 from dotenv import load_dotenv
 
 from openhands.core.config import load_app_config
-from openhands.server.config.openhands_config import load_openhands_config
+from openhands.server.config.server_config import load_server_config
 from openhands.server.conversation_manager.conversation_manager import (
     ConversationManager,
 )
@@ -14,7 +14,7 @@ from openhands.utils.import_utils import get_impl
 load_dotenv()
 
 config = load_app_config()
-openhands_config = load_openhands_config()
+server_config = load_server_config()
 file_store = get_file_store(config.file_store, config.file_store_path)
 
 client_manager = None
@@ -32,6 +32,6 @@ sio = socketio.AsyncServer(
 
 ConversationManagerImpl = get_impl(
     ConversationManager,  # type: ignore
-    openhands_config.conversation_manager_class,
+    server_config.conversation_manager_class,
 )
 conversation_manager = ConversationManagerImpl.get_instance(sio, config, file_store)
